* content() has new options to return the page formatted as text, with a       <base href> added. (RT#8087, patch by Dominique Quatravaux)
    * update_html() method has been added, which can be used to modify the HTML
      that Mech parses. It should be sub-classed instead of _parse_html(), which has
      been deprecated. (RT#8087, patch by Dominique Quatravaux)

####

I thought more about the implications of the new content() API forward compatibility.
What happens when we want to add options for other content types?

When the time comes, we can simple add a 'type => 'content' option, or better
auto-detect the content type and do something with that. I think we have some options
that look decent.

Mech is already centered around HTML, and I think sense to make the HTML-centric
options uncomplicated like these are.

I simply modified the docs to mention that the behavior for what content() does with
non-HTML content types /may/ change with options are present.