move LocalServer env kill earlier and add more variables
LocalServer loads LWP::Simple which creates a $ua so we have to clean 'em
out earlier. Plus this way even if the $mech is created first it's still
gone. Adding the https vars is probably unnecessary but I basically included
the full list from the most hostile proxied environment I had a user on.