rework set_field() to nth value implementation
This commit makes sure that we can still rely on $input->value
respecting $form->strict's state, which was broken (probably because it
didn't yet exist) in the original implementation of this feature.

Also adds more useful tests and updates the mailmap to make sure we
credit the original author correctly.