allow 2 empty consecutive lines in code
When we build our dist, POD::Weaver cuts out POD blocks on top of every
method and puts it all at the bottom of the file. But it leaves behind
the two empty lines of padding, which cannot be disabled in Weaver. To
mitigate that, we will allow having two empty lines. It might lead to
some code getting a bit wider, but the alternative solutions all seemed
much more badder.